Sam3s DAC PDC transfer now uses PWM for sampling signal generation.
One caveat for now: the sampling signal has to be twice as one would expect. I don't know why. All the registers seem right.

For future improvement: the frequency calculations should be all done in Hz. Right now it is a mix of kHz and Hz. This looses accuracy in frequency generation.