Whole second timer precision and multiplexed Alarms.
Adds TSecond as a timer resolution, with support for alarms, counters, and
LocalTime on msp430.  Past discussion on this feature can be found at
https://www.millennium.berkeley.edu/pipermail/tinyos-devel/2010-July/004494.html

Adds a multiplexed Alarm interface, much like how Timer is done, using the
existing VirtualizeAlarm capability.  In many network protocols, an alarm is
a better abstraction than a timer, and given the number of distinct alarms
that have to be maintained, virtualization is a requirement.